Auth Changes

+++ controllers/contactController.js
@@ -4,7 +4,7 @@ const Contact = require ("../models/contactModel")
 //@route GET /api/contacts
 
 const getContacts = asyncHandler( async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json({contacts});
 });
 
@@ -28,7 +28,7 @@ const createContact =asyncHandler( async (req, res) => {
   }
 
   const contact = await Contact.create({
-    name,email,phone,
+    name,email,phone, user_id: req.user.id
   })
   res.status(201).json({ contact});
 });
@@ -38,6 +38,11 @@ const updateContact =  asyncHandler( async (req, res) => {
   if (!contact){
     res.status(404);
     throw new Error ("Contact not found")
+  }
+
+  if (contact.user_id.toString () !== req.user.id){
+    res.status(403);
+    throw new Error ("User dont have the permission to update other user")
   }
    const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -53,6 +58,10 @@ const deleteContact =  asyncHandler( async (req, res) => {
   if (!contact){
     res.status(404);
     throw new Error ("Contact not found")
+  }
+  if (contact.user_id.toString () !== req.user.id){
+    res.status(403);
+    throw new Error ("User dont have the permission to delete other user")
   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json({ message: "Contact deleted successfully" });
